fix: structs are not packed in memory

+++ src/interfaces/ITerms.sol
@@ -39,7 +39,7 @@ struct Signature {
 }
 
 struct Price {
-    uint248 price;
+    uint256 price;
     bool seen;
 }
